refactor: AddressManager 컴포넌트 AddressSection 내에서 직접 구현하도록 변경
