+++ src/components/organisms/AddressSection/AddressSection.module.css
@@ -25,3 +25,49 @@
   height: 2.25rem;
   width: 6.25rem;
 }
+.wrapper {
+  display: flex;
+  flex-direction: column;
+  gap: 1rem;
+  background-color: var(--color-white);
+  border: 1px solid var(--color-gray-300);
+  border-radius: 0.5rem;
+  padding: 2rem;
+}
+.header {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+}
+.labelInputGroup {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+.grid2 {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 1rem;
+}
+.field {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+}
+.actions {
+  display: flex;
+  gap: 0.5rem;
+  margin-left: auto;
+}
+.actions button {
+  height: 2.25rem;
+  width: 3.6rem;
+}
+.type {
+  font-size: 1rem;
+  font-weight: 500;
+  color: var(--color-black);
+}
+.details {
+  color: var(--color-text);
+}
